# docs/yapi-ptz-biz-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "视频中台 - 云台监控 / 业务端",
    "description": "仅含云台监控与业务端（目录、设备、直播、录像、录像机回放、片段截取）。\n除目录树、设备详情、录像机回放、片段文件外，带参接口一律 POST + JSON Body。\n统一响应 { code, message, data }；code=0 成功。时间字段为毫秒时间戳。\n设备 status：0=不可用 1=已启用 2=已停用。直播仅 status=1；回放/截取 status=2 不可用。\n一台设备只对接一个平台 vendor：MOCK 模拟、UNIVIEW 宇视、HIKVISION 海康。调用方不传平台，只传 deviceId，服务按 vendor 调用对应平台。云台路径为 /api/ptz。海康尚未实现，失败文案为「海康尚未对接」。\n返回只有调用方会用到的字段。没有密码、经纬度、创建时间、服务器磁盘路径。\n宇视云台发给这台摄像机自己的地址；历史录像在它绑定的录像机上。示例设备 UV_10135。\n鉴权：auth.enabled=false 时免登录；true 时 Header Authorization: Bearer {token}。<video> 带不了 Header，URL 上加 token。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://8.130.74.232:8090",
      "description": "线上环境"
    },
    {
      "url": "http://127.0.0.1:8090",
      "description": "本地环境"
    }
  ],
...
  "paths": {
...
    "/api/biz/devices/list": {
      "post": {
        "tags": [
          "业务端"
        ],
        "summary": "业务端-设备列表",
        "description": "含已停用设备；用 playable、livePlayable 判断可否回放/直播。不返回码流明细。Body 可传 {}。",
        "operationId": "bizDevices",
        "requestBody": {
          "required": false,
          "description": "JSON 请求体",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BizDeviceListRequest"
              },
              "example": {
                "folderId": 2,
                "includeChildren": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功。失败时 HTTP 多为 400，body 中 code!=0，message 为原因，data 常为 null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespBizDeviceList"
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": [
                    {
                      "deviceId": "UV_10135",
                      "name": "演示球机",
                      "address": "宇视在线调试",
                      "folderId": 2,
                      "status": 1,
                      "playable": true,
                      "livePlayable": true
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/biz/devices/detail": {
      "post": {
        "tags": [
          "业务端"
        ],
        "summary": "业务端-设备详情",
        "description": "单台设备，并带上业务直播流地址。不拉流。没有业务直播流时 liveStream 为 null。不存在时 message=设备不存在: {deviceId}。",
        "operationId": "bizDeviceDetail",
        "requestBody": {
          "required": true,
          "description": "JSON 请求体",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BizDeviceIdRequest"
              },
              "example": {
                "deviceId": "UV_10135"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功。失败时 HTTP 多为 400，body 中 code!=0，message 为原因，data 常为 null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespBizDeviceDetail"
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "deviceId": "UV_10135",
                    "name": "演示球机",
                    "address": "宇视在线调试",
                    "folderId": 2,
                    "status": 1,
                    "playable": true,
                    "livePlayable": true,
                    "liveStream": {
                      "streamType": "sub",
                      "streamUrl": "http://8.130.74.232:8080/live/uv_UV_10135_sub.live.flv"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/biz/devices/recording-days": {
      "post": {
        "tags": [
          "业务端"
        ],
        "summary": "业务端-有录像的日期",
        "description": "给回放日历打点。宇视且已绑定录像机时查录像机；宇视未绑定以及模拟设备查中台本地 MP4。海康失败：海康尚未对接。已停用设备不能查。返回 yyyy-MM-dd，升序。",
        "operationId": "bizRecordingDays",
        "requestBody": {
          "required": true,
          "description": "JSON 请求体",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BizRecordingDaysRequest"
              },
              "example": {
                "deviceId": "UV_10135",
                "year": 2026,
                "month": 9
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功。失败时 HTTP 多为 400，body 中 code!=0，message 为原因，data 常为 null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespRecordingDays"
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": [
                    "2026-09-10",
                    "2026-09-15",
                    "2026-09-17"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/biz/devices/recordings": {
      "post": {
        "tags": [
          "业务端"
        ],
        "summary": "业务端-录像列表",
        "description": "数据源与有录像的日期相同：宇视且已绑定录像机查录像机，模拟或未绑定查本地 MP4，海康失败：海康尚未对接。已停用设备不能查。已绑定录像机的记录带 source=nvr，没有 videoUrl，播放走 GET /api/recordings/playback.flv。本地 MP4 带 durationSeconds 和 videoUrl。",
        "operationId": "bizRecordings",
        "requestBody": {
          "required": true,
          "description": "JSON 请求体",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BizRecordingsRequest"
              },
              "example": {
                "deviceId": "UV_10135",
                "from": 1726502400000,
                "to": 1726588799999
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功。失败时 HTTP 多为 400，body 中 code!=0，message 为原因，data 常为 null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespRecordingUrlList"
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": [
                    {
                      "deviceId": "UV_10135",
                      "fileName": "C1/B1790092800/E1790093684",
                      "recordTime": 1790092800000,
                      "endTime": 1790093684000,
                      "size": 0,
                      "source": "nvr"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/biz/devices/{deviceId}/clip": {
      "get": {
        "tags": [
          "业务端"
        ],
        "summary": "业务端-按时间点截取",
        "description": "以 at 为中心，前后各 seconds 秒切出一段。本接口只查录像并返回播放地址，通常几秒内返回，不等待文件生成。宇视且已绑定录像机时地址指向录像机历史；模拟或未绑定指向本地 MP4。海康失败：海康尚未对接。MP4 在访问 videoUrl 时才生成，耗时约等于片段时长。该时间点没有对应录像：该时间点无可用录像。",
        "operationId": "bizClip",
        "parameters": [
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "description": "设备编码",
            "schema": {
              "type": "string",
              "description": "设备编码",
              "example": "UV_10135"
            },
            "example": "UV_10135"
          },
          {
            "name": "at",
            "in": "query",
            "required": true,
            "description": "时间点。10 位是秒，11~13 位是毫秒",
            "schema": {
              "type": "string",
              "description": "时间点。10 位是秒，11~13 位是毫秒",
              "example": "1730000030000"
            },
            "example": "1730000030000"
          },
          {
            "name": "seconds",
            "in": "query",
            "required": false,
            "description": "前后各取的秒数，默认 30，须大于 0，最大 150",
            "schema": {
              "type": "integer",
              "description": "前后各取的秒数，默认 30，须大于 0，最大 150",
              "example": 30
            },
            "example": 30
          }
        ],
        "responses": {
          "200": {
            "description": "成功。失败时 HTTP 多为 400，body 中 code!=0，message 为原因，data 常为 null",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "统一 JSON 响应：code / message / data",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "业务状态码：0=成功，非 0=失败（一般为 -1）",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "description": "提示信息：成功一般为 success，失败为错误原因",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ClipInfo"
                    }
                  },
                  "example": {
                    "code": 0,
                    "message": "success",
                    "data": {
                      "deviceId": "UV_10135",
                      "at": 1730000030000,
                      "seconds": 30,
                      "startTime": 1730000000000,
                      "endTime": 1730000060000,
                      "durationSeconds": 60,
                      "videoUrl": "http://8.130.74.232:8090/api/biz/devices/UV_10135/clip/file?at=1730000030000&seconds=30",
                      "clipUrl": "http://8.130.74.232:8090/api/biz/devices/UV_10135/clip/file?at=1730000030000&seconds=30",
                      "windowStart": 1730000000000,
                      "windowEnd": 1730000060000,
                      "sourceFiles": [
                        "20260910_143000.mp4"
                      ],
                      "clipFileName": "UV_10135_1730000030000_30.mp4",
                      "size": 5242880,
                      "ok": true
                    }
                  }
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "deviceId": "UV_10135",
                    "at": 1730000030000,
                    "seconds": 30,
                    "startTime": 1730000000000,
                    "endTime": 1730000060000,
                    "durationSeconds": 60,
                    "videoUrl": "http://8.130.74.232:8090/api/biz/devices/UV_10135/clip/file?at=1730000030000&seconds=30",
                    "clipUrl": "http://8.130.74.232:8090/api/biz/devices/UV_10135/clip/file?at=1730000030000&seconds=30",
                    "windowStart": 1730000000000,
                    "windowEnd": 1730000060000,
                    "sourceFiles": [
                      "20260910_143000.mp4"
                    ],
                    "clipFileName": "UV_10135_1730000030000_30.mp4",
                    "size": 5242880
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/biz/clips": {
      "post": {
        "tags": [
          "业务端"
        ],
        "summary": "业务端-批量截取录像片段",
        "description": "一次切多段。每条按该设备的平台取录像，规则与按时间点截取相同。Body 是 JSON 数组，不是 {items:[]}。最多 50 条。单条失败不影响其它条。成功条带截取结果字段且 ok=true；失败条 ok=false，并有 deviceId、at、seconds、error。海康那一条 error 为海康尚未对接。",
        "operationId": "bizClipsBatch",
        "requestBody": {
          "required": true,
          "description": "JSON 请求体",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "description": "批量截取请求",
                "items": {
                  "$ref": "#/components/schemas/RecordClipItemRequest"
                }
              },
              "example": [
                {
                  "deviceId": "UV_10135",
                  "at": 1730000030000,
                  "seconds": 30
                },
                {
                  "deviceId": "CAM_WEST_02",
                  "at": 1730000100000,
                  "seconds": 15
                }
              ]
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功。失败时 HTTP 多为 400，body 中 code!=0，message 为原因，data 常为 null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespClipBatch"
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": [
                    {
                      "deviceId": "UV_10135",
                      "at": 1730000030000,
                      "seconds": 30,
                      "startTime": 1730000000000,
                      "endTime": 1730000060000,
                      "durationSeconds": 60,
                      "videoUrl": "http://8.130.74.232:8090/api/biz/devices/UV_10135/clip/file?at=1730000030000&seconds=30",
                      "clipUrl": "http://8.130.74.232:8090/api/biz/devices/UV_10135/clip/file?at=1730000030000&seconds=30",
                      "windowStart": 1730000000000,
                      "windowEnd": 1730000060000,
                      "sourceFiles": [
                        "20260910_143000.mp4"
                      ],
                      "clipFileName": "UV_10135_1730000030000_30.mp4",
                      "size": 5242880,
                      "ok": true
                    },
                    {
                      "deviceId": "CAM_WEST_02",
                      "at": 1730000100000,
                      "seconds": 15,
                      "ok": false,
                      "error": "该时间点无可用录像"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/biz/devices/{deviceId}/clip/file": {
      "get": {
        "tags": [
          "业务端"
        ],
        "summary": "业务端-片段文件",
        "description": "按时间点截取和批量截取返回的 videoUrl。第一次访问才生成 MP4，耗时约等于片段时长；生成后直接返回缓存。响应是 MP4，不是 JSON。",
        "operationId": "bizClipFile",
        "parameters": [
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "description": "设备编码",
            "schema": {
              "type": "string",
              "description": "设备编码",
              "example": "UV_10135"
            },
            "example": "UV_10135"
          },
          {
            "name": "at",
            "in": "query",
            "required": true,
            "description": "与截取时相同的时间点",
            "schema": {
              "type": "string",
              "description": "与截取时相同的时间点",
              "example": "1730000030000"
            },
            "example": "1730000030000"
          },
          {
            "name": "seconds",
            "in": "query",
            "required": false,
            "description": "与截取时相同，默认 30",
            "schema": {
              "type": "integer",
              "description": "与截取时相同，默认 30",
              "example": 30
            },
            "example": 30
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "description": "开启鉴权时给 video 标签用",
            "schema": {
              "type": "string",
              "description": "开启鉴权时给 video 标签用"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "MP4 文件流",
            "content": {
              "video/mp4": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ClipInfo": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "description": "设备编码",
            "example": "UV_10135"
          },
          "at": {
            "type": "integer",
            "description": "时间点，已换成毫秒",
            "example": 1730000030000,
            "format": "int64"
          },
          "seconds": {
            "type": "integer",
            "description": "实际使用的前后秒数",
            "example": 30
          },
          "windowStart": {
            "type": "integer",
            "description": "请求窗口起点，at − seconds×1000",
            "example": 1726560000000,
            "format": "int64"
          },
          "windowEnd": {
            "type": "integer",
            "description": "请求窗口终点，at + seconds×1000",
            "example": 1726560000000,
            "format": "int64"
          },
          "startTime": {
            "type": "integer",
            "description": "片段实际开始，受录像覆盖裁切",
            "example": 1726560000000,
            "format": "int64"
          },
          "endTime": {
            "type": "integer",
            "description": "片段实际结束",
            "example": 1726560000000,
            "format": "int64"
          },
          "durationSeconds": {
            "type": "number",
            "description": "实际时长秒",
            "example": 60
          },
          "videoUrl": {
            "type": "string",
            "description": "给播放器的 MP4 地址，即片段文件接口",
            "example": "http://8.130.74.232:8090/api/biz/devices/UV_10135/clip/file?at=1730000030000&seconds=30"
          },
          "clipUrl": {
            "type": "string",
            "description": "与 videoUrl 相同",
            "example": "http://8.130.74.232:8090/api/biz/devices/UV_10135/clip/file?at=1730000030000&seconds=30"
          },
          "sourceFiles": {
            "type": "array",
            "description": "源文件名",
            "items": {
              "type": "string",
              "description": "参与截取的原始文件名"
            }
          },
          "clipFileName": {
            "type": "string",
            "description": "缓存片段文件名",
            "example": "UV_10135_1730000030000_30.mp4"
          },
          "size": {
            "type": "integer",
            "description": "片段字节数。文件尚未生成时不返回",
            "example": 5242880,
            "nullable": true,
            "format": "int64"
          }
        },
        "description": "按时间点截取的结果。宇视且已绑定录像机时来自录像机，否则来自本地 MP4"
      },
...
      "RespBizDeviceList": {
        "type": "object",
        "description": "统一 JSON 响应：code / message / data",
        "properties": {
          "code": {
            "type": "integer",
            "description": "业务状态码：0=成功，非 0=失败（一般为 -1）",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示信息：成功一般为 success，失败为错误原因",
            "example": "success"
          },
          "data": {
            "type": "array",
            "description": "业务端设备列表",
            "items": {
              "$ref": "#/components/schemas/BizDeviceItem"
            }
          }
        },
        "example": {
          "code": 0,
          "message": "success",
          "data": [
            {
              "deviceId": "UV_10135",
              "name": "演示球机",
              "address": "宇视在线调试",
              "folderId": 2,
              "status": 1,
              "playable": true,
              "livePlayable": true
            }
          ]
        }
      },
      "RespBizDeviceDetail": {
        "type": "object",
        "description": "统一 JSON 响应：code / message / data",
        "properties": {
          "code": {
            "type": "integer",
            "description": "业务状态码：0=成功，非 0=失败（一般为 -1）",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示信息：成功一般为 success，失败为错误原因",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/BizDeviceDetail"
          }
        },
        "example": {
          "code": 0,
          "message": "success",
          "data": {
            "deviceId": "UV_10135",
            "name": "演示球机",
            "address": "宇视在线调试",
            "folderId": 2,
            "status": 1,
            "playable": true,
            "livePlayable": true,
            "liveStream": {
              "streamType": "sub",
              "streamUrl": "http://8.130.74.232:8080/live/uv_UV_10135_sub.live.flv"
            }
          }
        }
      },
...
      "RespRecordingDays": {
        "type": "object",
        "description": "统一 JSON 响应：code / message / data",
        "properties": {
          "code": {
            "type": "integer",
            "description": "业务状态码：0=成功，非 0=失败（一般为 -1）",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示信息：成功一般为 success，失败为错误原因",
            "example": "success"
          },
          "data": {
            "type": "array",
            "description": "某月内有录像的日期列表（升序）",
            "items": {
              "type": "string",
              "description": "有录像的日期 yyyy-MM-dd",
              "example": "2026-09-17"
            }
          }
        },
        "example": {
          "code": 0,
          "message": "success",
          "data": [
            "2026-09-10",
            "2026-09-15",
            "2026-09-17"
          ]
        }
      },
      "RespRecordingUrlList": {
        "type": "object",
        "description": "统一 JSON 响应：code / message / data",
        "properties": {
          "code": {
            "type": "integer",
            "description": "业务状态码：0=成功，非 0=失败（一般为 -1）",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示信息：成功一般为 success，失败为错误原因",
            "example": "success"
          },
          "data": {
            "type": "array",
            "description": "录像列表（含 videoUrl）",
            "items": {
              "$ref": "#/components/schemas/RecordingWithUrl"
            }
          }
        },
        "example": {
          "code": 0,
          "message": "success",
          "data": [
            {
              "deviceId": "UV_10135",
              "fileName": "C1/B1790092800/E1790093684",
              "recordTime": 1790092800000,
              "endTime": 1790093684000,
              "size": 0,
              "source": "nvr"
            }
          ]
        }
      },
...
      "BizDeviceListRequest": {
        "type": "object",
        "properties": {
          "folderId": {
            "type": "integer",
            "description": "按目录过滤：目录主键 ID；不传则返回全部设备",
            "example": 2,
            "nullable": true,
            "format": "int64"
          },
          "includeChildren": {
            "type": "boolean",
            "description": "当传入 folderId 时：true=包含子孙（默认 true）；false=仅本目录",
            "example": true
          }
        },
        "description": "业务端设备列表请求；可传空对象 {}"
      },
      "BizDeviceIdRequest": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "description": "【必填】设备业务编码",
            "example": "UV_10135"
          }
        },
        "required": [
          "deviceId"
        ],
        "description": "按设备编码查询/操作"
      },
      "BizRecordingDaysRequest": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "description": "【必填】设备业务编码",
            "example": "UV_10135"
          },
          "year": {
            "type": "integer",
            "description": "【必填】年份 1970~2100",
            "example": 2026
          },
          "month": {
            "type": "integer",
            "description": "【必填】月份 1~12",
            "example": 9
          }
        },
        "required": [
          "deviceId",
          "year",
          "month"
        ],
        "description": "查询某月有录像的日期"
      },
      "BizRecordingsRequest": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "description": "【必填】设备业务编码",
            "example": "UV_10135"
          },
          "from": {
            "type": "integer",
            "description": "开始时间（含），毫秒时间戳；不传则不限起始",
            "example": 1726502400000,
            "nullable": true,
            "format": "int64"
          },
          "to": {
            "type": "integer",
            "description": "结束时间（含），毫秒时间戳；不传则不限结束",
            "example": 1726588799999,
            "nullable": true,
            "format": "int64"
          }
        },
        "required": [
          "deviceId"
        ],
        "description": "录像列表请求"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "auth.enabled=true 时需要。登录后填 JWT，不要带 Bearer 前缀（YApi 会加）。"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}